requirements.txt
# libraries
openai==1.30.2
python-telegram-bot==21.2
python-dotenv==1.0.1
jupyter==1.0.0
loguru==0.7.2
